+++ index.html
@@ -54,9 +54,7 @@
 
   <div class="container mb-5">
     <div id="div1" class="d-flex flex-row justify-content-center">
-      <span class="time">09</span>
-      <span class="time">:</span>
-      <span class="time">47</span>
+      <span id="m-time" class="time"></span>
     </div>
   </div>
 
@@ -125,7 +123,8 @@
     <!-- Copyright -->
   </footer>
 
-  <script type="module" src="index.js"></script>
+  <script type="module" src="./index.js"></script>
+  <script src="./modules/bookList.js"></script>
   <script src="https://moment.github.io/luxon/global/luxon.js"></script>
