+++ index.js
@@ -1,0 +1,9 @@
+import { date } from './modules/date.js';
+// import bookList from './modules/bookList.js';
+// import { addBook } from './modules/addBook.js';
+// import { removeBook } from './modules/removeBook.js';
+
+date();
+// bookList = new BookList();
+// addBook();
+// removeBook();
