+++ modules/addbook.js
@@ -1,84 +1,28 @@
-class BookList {
-    constructor() {
-        this.bookArr = [];
+// // Submit Event Listener
+// export function addBook(bookAuthor, bookTitle, bookIsbn) {
+//     const book = {
+//         title: bookTitle,
+//         author: bookAuthor,
+//         isbn: bookIsbn,
+//     };
 
-        // check if books are already stored in the localStorage
-        if (localStorage.getItem('books') !== null) {
-            // collect array of books from local storage
-            this.bookArr = JSON.parse(localStorage.getItem('books'));
+//     this.bookArr.push(book);
 
-            // loop through books to display them to browser
-            for (let i = 0; i < this.bookArr.length; i += 1) {
-                this.displayBook(this.bookArr[i].isbn);
-            }
-        }
-    }
+//     this.displayBook(book.isbn);
 
-    addBook(bookAuthor, bookTitle, bookIsbn) {
-        const book = {
-            title: bookTitle,
-            author: bookAuthor,
-            isbn: bookIsbn,
-        };
+//     localStorage.setItem('books', JSON.stringify(this.bookArr));
+// }
 
-        this.bookArr.push(book);
+// document.querySelector('#book-form').addEventListener('submit', (e) => {
+//     e.preventDefault();
 
-        this.displayBook(book.isbn);
+//     const author = document.querySelector('#author').value;
+//     const title = document.querySelector('#title').value;
+//     const isbn = document.querySelector('#isbn').value;
 
-        localStorage.setItem('books', JSON.stringify(this.bookArr));
-    }
+//     bookList.addBook(author, title, isbn);
 
-    displayBook(isbn) {
-        let book;
-
-        // loop through books to display them to browser
-        for (let i = 0; i < this.bookArr.length; i += 1) {
-            if (this.bookArr[i].isbn === isbn) {
-                book = this.bookArr[i];
-            }
-        }
-
-        const tr = document.createElement('tr');
-        tr.setAttribute('data-isbn', book.isbn);
-        tr.innerHTML = `<td>${book.title}</td><td>${book.author}</td><td>${book.isbn}</td><td class = ""><a href="#" class = "btn btn-sm btn-danger delete">X</a></td>`;
-
-        document.querySelector('#book-list').appendChild(tr);
-    }
-
-    removeBook(isbn) {
-        this.bookArr = this.bookArr.filter((book) => book.isbn !== isbn);
-
-        localStorage.setItem('books', JSON.stringify(this.bookArr));
-    }
-}
-
-const bookList = new BookList();
-
-// Submit Event Listener
-document.querySelector('#book-form').addEventListener('submit', (e) => {
-    e.preventDefault();
-
-    const author = document.querySelector('#author').value;
-    const title = document.querySelector('#title').value;
-    const isbn = document.querySelector('#isbn').value;
-
-    bookList.addBook(author, title, isbn);
-
-    document.querySelector('#author').value = '';
-    document.querySelector('#title').value = '';
-    document.querySelector('#isbn').value = '';
-});
-
-// Event listener when the user clicks on remove button
-document.addEventListener('click', (e) => {
-    const deleteButton = e.target;
-
-    if (deleteButton !== undefined && deleteButton.classList.contains('delete')) {
-        e.preventDefault();
-        const tr = deleteButton.closest('tr');
-        const isbn = tr.getAttribute('data-isbn');
-        tr.remove();
-
-        bookList.removeBook(isbn);
-    }
-});
+//     document.querySelector('#author').value = '';
+//     document.querySelector('#title').value = '';
+//     document.querySelector('#isbn').value = '';
+// });
