+++ modules/date.js
@@ -1,0 +1,12 @@
+const myDate = document.getElementById('m-time');
+export const date = () => {
+    window.setInterval(
+        () => {
+            myDate.innerHTML = `<p>${luxon.DateTime.local().toLocaleString(
+                luxon.DateTime.DATETIME_MED_WITH_SECONDS,
+            )}</p>`;
+        },
+        1000,
+        this,
+    );
+};
